Fully implemented image resizing.
First of all - WebImage is not used anymore, as sometimes it leaves a
1px black border after resizing. (see
http://aspnetwebstack.codeplex.com/workitem/1126 ).

New resizing code is borrowed from
http://stackoverflow.com/questions/87753/resizing-an-image-without-losing-any-quality
with some fixes and modifications.
Actually I'm not quite sure it's a best (or even good) solution, but I
didn't want to introduce any external image resizing libraries as
dependencies so here it goes.

If anyone comes up with a better solution, it will be greatly
appreciated.